pyproject.toml
[project]
name = "portal-ldap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "python-ldap>=3.4.4",
]

[dependency-groups]
dev = [
    "ruff>=0.12.5",
]
